requirements.txt
fastapi
gunicorn
uvicorn
jinja2
pytest
requests
pre-commit
aiofiles
python-multipart
pillow
pytesseract